# DAML Engine: Speed up foldl